Alpha merge

+++ modules/saverestore/saverestore.class.php
@@ -396,12 +396,10 @@ class saverestore extends module
 
         if ($this->mode == 'upload') {
             $this->upload($out);
-            //$this->redirect("?mode=clear");
         }
         if ($this->mode == 'dump') {
             $this->dump($out);
             $this->redirect("?mode=clear");
-            //$this->redirect("?");
         }
 
         if ($this->mode == 'delete') {
@@ -1268,7 +1266,6 @@ class saverestore extends module
      */
     function upload(&$out, $iframe = 0)
     {
-
         set_time_limit(0);
         global $restore;
         global $file;
@@ -1291,7 +1288,6 @@ class saverestore extends module
             move_uploaded_file($file, DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/' . $file_name);
             $file = $file_name;
         }
-
 
         if ($iframe) {
             echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_APPLY_UPDATE . '</div>');
@@ -1335,6 +1331,9 @@ class saverestore extends module
                 $this->redirect("?mode=clear&ok_msg=" . urlencode(LANG_UPDATEBACKUP_RESTORE_DB_DONE));
             }
         } elseif ($file != '') {
+
+            DebMes("Trying to unpack $file", "restore");
+
             logAction('system_restore', $file);
             // unpack archive
             umask(0);
@@ -1349,14 +1348,25 @@ class saverestore extends module
             } else {
                 $result = exec('tar xzvf ../' . $file, $output, $res);
             }
+
             if (!$result) {
                 echonow("Unpack failed", 'red');
+                DebMes("Unpack failed: " . 'tar xzvf ../' . $file, "restore");
                 return false;
             }
 
 
-            $UpdatesDir = scandir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp', 1);
-            $folder = DIRECTORY_SEPARATOR . $UpdatesDir[0];
+            if (file_exists(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/index.php') || file_exists(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/dump.sql')) {
+                $folder = '/.';
+            } else {
+                $UpdatesDir = scandir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp', 1);
+                $folder = DIRECTORY_SEPARATOR . $UpdatesDir[0];
+                if (!is_dir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder)) {
+                    echonow("Unpack failed!", 'red');
+                    return false;
+                }
+            }
+            DebMes("Restore folder: $folder", "restore");
 
             if ($iframe) {
                 echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_DONE . '</div>');
@@ -1392,11 +1402,11 @@ class saverestore extends module
             $total = count($ignores);
             for ($i = 0; $i < $total; $i++) {
                 $name = $ignores[$i]['NAME'];
-                if (is_dir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/modules/' . $name)) {
-                    removeTree(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/modules/' . $name);
+                if (is_dir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/modules/' . $name)) {
+                    removeTree(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/modules/' . $name);
                 }
-                if (is_dir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/templates/' . $name)) {
-                    removeTree(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp/templates/' . $name);
+                if (is_dir(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/templates/' . $name)) {
+                    removeTree(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/templates/' . $name);
                 }
             }
 
@@ -1420,8 +1430,10 @@ class saverestore extends module
                 if ($iframe) {
                     echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_RESTORE_DB . '</div>');
                 }
-                $this->restoredatabase(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/dump.sql');
-                echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_DONE . '</div>');
+                $result = $this->restoredatabase(DOC_ROOT . DIRECTORY_SEPARATOR . 'cms/saverestore/temp' . $folder . '/dump.sql');
+                if ($iframe) {
+                    echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_DONE . '</div>');
+                }
             }
 
             $this->config['LATEST_UPDATED_ID'] = $out['LATEST_ID'];
@@ -1436,7 +1448,6 @@ class saverestore extends module
             if ($iframe) {
                 echonow('<div><i style="font-size: 7pt;" class="glyphicon glyphicon-usd"></i> ' . LANG_UPDATEBACKUP_DONE . '</div>');
             }
-
 
             if ($iframe) {
                 return 1;
@@ -1623,11 +1634,19 @@ class saverestore extends module
     {
         $mysql_path = (substr(php_uname(), 0, 7) == "Windows") ? SERVER_ROOT . "/server/mysql/bin/mysql" : 'mysql';
         $mysqlParam = " -u " . DB_USER;
-        if (DB_PASSWORD != '') $mysqlParam .= " -p" . DB_PASSWORD;
+        if (DB_PASSWORD != '') $mysqlParam .= " --password=\"" . DB_PASSWORD . "\"";
         $mysqlParam .= " " . DB_NAME . " <" . $filename;
-        exec($mysql_path . $mysqlParam);
-        SQLExec("DELETE FROM cached_values");
-        setGlobal('cycle_mainRun', time());
+        $cmd = $mysql_path . $mysqlParam;
+        $result = exec($cmd, $output, $result_code);
+        if ($result !== false) {
+            DebMes("DB restored", "restore");
+            SQLExec("DELETE FROM cached_values");
+            setGlobal('cycle_mainRun', time());
+            return true;
+        } else {
+            DebMes("Failed to restore DB:\n" . implode("\n", $output), "restore");
+            return false;
+        }
     }
 
     /**
